Refactor testing constants
