+++ test/datasources.ts
@@ -7,6 +7,7 @@ import {
   AccelerationsCacheData,
   AssociatedObject,
   AsyncPollingResult,
+  CachedAcceleration,
   CachedDataSourceStatus,
   DataSourceCacheData,
   DatasourceDetails,
@@ -1199,3 +1200,33 @@ export const mockAssociatedObjects: AssociatedObject[] = [
     ],
   },
 ];
+
+export const skippingIndexAcceleration = {
+  flintIndexName: 'flint_mys3_default_http_logs_skipping_index',
+  type: 'skipping',
+  database: 'default',
+  table: 'http_logs',
+  indexName: '',
+  autoRefresh: false,
+  status: 'active',
+} as CachedAcceleration;
+
+export const materializedViewAcceleration = {
+  flintIndexName: 'flint_mys3_default_http_count_view',
+  type: 'materialized',
+  database: 'default',
+  table: '',
+  indexName: 'http_count_view',
+  autoRefresh: false,
+  status: 'active',
+} as CachedAcceleration;
+
+export const coveringIndexAcceleration = {
+  flintIndexName: 'flint_mys3_default_http_logs_status_clientip_and_day_index',
+  type: 'covering',
+  database: 'default',
+  table: 'http_logs',
+  indexName: 'status_clientip_and_day',
+  autoRefresh: true,
+  status: 'refreshing',
+} as CachedAcceleration;
